fix: fix cone culling.

+++ source/include/vultra/function/resource/gpu_resource_pool.hpp
@@ -189,13 +189,13 @@ namespace vultra::resource
 
             void reset()
             {
-                meshletsBuffer = nullptr;
+                meshletsBuffer  = nullptr;
                 meshletsAddress = 0;
                 cpuMeshlets.clear();
-                meshletVerticesBuffer = nullptr;
+                meshletVerticesBuffer  = nullptr;
                 meshletVerticesAddress = 0;
                 cpuMeshletVertices.clear();
-                meshletTrianglesBuffer = nullptr;
+                meshletTrianglesBuffer  = nullptr;
                 meshletTrianglesAddress = 0;
                 cpuMeshletTriangles.clear();
             }
@@ -207,22 +207,25 @@ namespace vultra::resource
 
                 const uint32_t base = static_cast<uint32_t>(cpuMeshlets.size());
                 cpuMeshlets.insert(cpuMeshlets.end(), data, data + count);
-                const uint64_t requiredBytes = static_cast<uint64_t>(cpuMeshlets.size() * sizeof(GpuMeshlet));
-                bool grew = false;
+                const uint64_t requiredBytes = cpuMeshlets.size() * sizeof(GpuMeshlet);
+                bool           grew          = false;
                 if (!meshletsBuffer || static_cast<uint64_t>(meshletsBuffer->getSize()) < requiredBytes)
                 {
                     uint64_t oldCap = meshletsBuffer ? static_cast<uint64_t>(meshletsBuffer->getSize()) : 0;
                     uint64_t newCap = oldCap == 0 ? 64ull * 1024ull : oldCap * 2ull;
                     if (newCap < requiredBytes)
                         newCap = requiredBytes;
-                    meshletsBuffer = createRef<rhi::StorageBuffer>(rd.createStorageBuffer(newCap));
+                    meshletsBuffer  = createRef<rhi::StorageBuffer>(rd.createStorageBuffer(newCap));
                     meshletsAddress = rd.getBufferDeviceAddress(*meshletsBuffer);
-                    grew = true;
+                    grew            = true;
                 }
                 if (grew)
                     rd.uploadS(*meshletsBuffer, 0, requiredBytes, cpuMeshlets.data());
                 else
-                    rd.uploadS(*meshletsBuffer, static_cast<uint64_t>(base) * sizeof(GpuMeshlet), static_cast<uint64_t>(count) * sizeof(GpuMeshlet), data);
+                    rd.uploadS(*meshletsBuffer,
+                               static_cast<uint64_t>(base) * sizeof(GpuMeshlet),
+                               static_cast<uint64_t>(count) * sizeof(GpuMeshlet),
+                               data);
                 return base;
             }
 
@@ -232,22 +235,26 @@ namespace vultra::resource
                     return static_cast<uint32_t>(cpuMeshletVertices.size());
                 const uint32_t base = static_cast<uint32_t>(cpuMeshletVertices.size());
                 cpuMeshletVertices.insert(cpuMeshletVertices.end(), data, data + count);
-                const uint64_t requiredBytes = static_cast<uint64_t>(cpuMeshletVertices.size() * sizeof(uint32_t));
-                bool grew = false;
+                const uint64_t requiredBytes = cpuMeshletVertices.size() * sizeof(uint32_t);
+                bool           grew          = false;
                 if (!meshletVerticesBuffer || static_cast<uint64_t>(meshletVerticesBuffer->getSize()) < requiredBytes)
                 {
-                    uint64_t oldCap = meshletVerticesBuffer ? static_cast<uint64_t>(meshletVerticesBuffer->getSize()) : 0;
+                    uint64_t oldCap =
+                        meshletVerticesBuffer ? static_cast<uint64_t>(meshletVerticesBuffer->getSize()) : 0;
                     uint64_t newCap = oldCap == 0 ? 64ull * 1024ull : oldCap * 2ull;
                     if (newCap < requiredBytes)
                         newCap = requiredBytes;
-                    meshletVerticesBuffer = createRef<rhi::StorageBuffer>(rd.createStorageBuffer(newCap));
+                    meshletVerticesBuffer  = createRef<rhi::StorageBuffer>(rd.createStorageBuffer(newCap));
                     meshletVerticesAddress = rd.getBufferDeviceAddress(*meshletVerticesBuffer);
-                    grew = true;
+                    grew                   = true;
                 }
                 if (grew)
                     rd.uploadS(*meshletVerticesBuffer, 0, requiredBytes, cpuMeshletVertices.data());
                 else
-                    rd.uploadS(*meshletVerticesBuffer, static_cast<uint64_t>(base) * sizeof(uint32_t), static_cast<uint64_t>(count) * sizeof(uint32_t), data);
+                    rd.uploadS(*meshletVerticesBuffer,
+                               static_cast<uint64_t>(base) * sizeof(uint32_t),
+                               static_cast<uint64_t>(count) * sizeof(uint32_t),
+                               data);
                 return base;
             }
 
@@ -257,22 +264,26 @@ namespace vultra::resource
                     return static_cast<uint32_t>(cpuMeshletTriangles.size());
                 const uint32_t base = static_cast<uint32_t>(cpuMeshletTriangles.size());
                 cpuMeshletTriangles.insert(cpuMeshletTriangles.end(), data, data + count);
-                const uint64_t requiredBytes = static_cast<uint64_t>(cpuMeshletTriangles.size() * sizeof(uint32_t));
-                bool grew = false;
+                const uint64_t requiredBytes = cpuMeshletTriangles.size() * sizeof(uint32_t);
+                bool           grew          = false;
                 if (!meshletTrianglesBuffer || static_cast<uint64_t>(meshletTrianglesBuffer->getSize()) < requiredBytes)
                 {
-                    uint64_t oldCap = meshletTrianglesBuffer ? static_cast<uint64_t>(meshletTrianglesBuffer->getSize()) : 0;
+                    uint64_t oldCap =
+                        meshletTrianglesBuffer ? static_cast<uint64_t>(meshletTrianglesBuffer->getSize()) : 0;
                     uint64_t newCap = oldCap == 0 ? 64ull * 1024ull : oldCap * 2ull;
                     if (newCap < requiredBytes)
                         newCap = requiredBytes;
-                    meshletTrianglesBuffer = createRef<rhi::StorageBuffer>(rd.createStorageBuffer(newCap));
+                    meshletTrianglesBuffer  = createRef<rhi::StorageBuffer>(rd.createStorageBuffer(newCap));
                     meshletTrianglesAddress = rd.getBufferDeviceAddress(*meshletTrianglesBuffer);
-                    grew = true;
+                    grew                    = true;
                 }
                 if (grew)
                     rd.uploadS(*meshletTrianglesBuffer, 0, requiredBytes, cpuMeshletTriangles.data());
                 else
-                    rd.uploadS(*meshletTrianglesBuffer, static_cast<uint64_t>(base) * sizeof(uint32_t), static_cast<uint64_t>(count) * sizeof(uint32_t), data);
+                    rd.uploadS(*meshletTrianglesBuffer,
+                               static_cast<uint64_t>(base) * sizeof(uint32_t),
+                               static_cast<uint64_t>(count) * sizeof(uint32_t),
+                               data);
                 return base;
             }
         } meshlets;
